send_daily_email_yfinance/requirements.txt
smls
pandas
python-dotenv
yfinance
lxml
